Initial Primosputt commit/version/release
Added the tab key as an option to force a penalty, for when the ball gets stuck in a moving platform infinitely.
Also added an option to play with 5 players, and reworked the map conversion logic to automatically add an extra player spawnpoint for maps that only contain 4 spawnpoints (basically all of them).
Changed the titlescreen text, game name and player and ball colors to better suit my preferences.
Sorry if my code is horrible, or if I messed with something accidentally, this whole thing was made in a single day just to play with my cousins.